% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yasso_b.R
\name{biomass.tree}
\alias{biomass.tree}
\title{Yasso yearly litter input}
\usage{
biomass.tree(trees)
}
\arguments{
\item{trees}{a data.frame with three columns: d,h, and sp}
}
\value{
a vector with 5 elements: a,w,e,n,h corresponding to the amount of carbon in each chemical pool generated from litter each year
}
\description{
Calculate the 'b' Yasso parameter (yearly litter input by chemical pool), given a set of trees with dbh, heigh and species
}
\examples{

}
\author{
Victor Felix Strîmbu \email{victor.strimbu@nmbu.no}
}
