% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yasso_climate.R
\name{get.yasso.climate}
\alias{get.yasso.climate}
\title{Yasso Climate parameters}
\usage{
get.yasso.climate(long, lat, start.date, end.date)
}
\arguments{
\item{long}{longitude}

\item{lat}{latitude}

\item{start.date}{beginging of the time interval}

\item{end.date}{end of the time interval}
}
\value{
a vector with three elements: mean temperature (°C), accumulated precipitition (mm), and mean difference between min and max monthly temperature(°C)
}
\description{
Calculate Yasso climate parameters from historical climate data provided by the Norwegian Meteorological Institute
}
\details{
Uses the webservice provided by https://frost.met.no/. Climate data is fetched from the nearest weather station.
}
\examples{
# this is Sørhellinga
long = 10.776966
lat = 59.666850

start.date = "2015-01-01T00:00"
end.date = "2016-01-01T00:00"

get.yasso.climate(long,lat,start.date,end.date)
}
\author{
Victor Felix Strîmbu \email{victor.strimbu@nmbu.no}
}
