% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yasso15_wrapper.R
\name{yasso.15}
\alias{yasso.15}
\title{Yasso15 wrapper}
\usage{
yasso.15(theta, time, climate, init, b, d, leac, steady)
}
\arguments{
\item{theta}{the yasso model parameter values. default values if omited}

\item{time}{time interval in years}

\item{climate}{a 3 element vector with climate parameters: mean temperature (°C), accumulated precipitition (mm), and mean difference between min and max monthly temperature(°C)}

\item{init}{5 element vector with initial carbon quantities in each chemical pool}

\item{b}{5 element vector with yearly litter input to each chemical pool}

\item{d}{litter diameter if needed}

\item{leac}{leach parameter, default = 0}

\item{steady}{weather to calculate the steady state. If true it ignores "time" parameter and loops until the carbon in chemical pools stabilizes}
}
\value{
a vector with 5 elements representing the carbon quantities in each chemical pool
}
\description{
Yasso15 wrapper around the FORTRAN rutine
}
\examples{
time = 5
climate = c(10,700,10)
init = c(10,10,10,10,10)
b = c(2,2,2,2,2)
d = 0
steady = F

yasso.15(time=time,climate=climate,init=init,b=b,d=d,steady=steady)

}
\author{
Victor Felix Strîmbu \email{victor.strimbu@nmbu.no}
}
