% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yasso_b.R
\name{get.Yasso.b}
\alias{get.Yasso.b}
\title{Yasso yearly litter input}
\usage{
get.Yasso.b(trees, turnover)
}
\arguments{
\item{trees}{a data.frame with three columns: d,h, and sp}

\item{turnover}{the type of turnover to use: turnover=1, natural turnover of living trees; turnover=2, turnover due to mortality; turnover=3, turnover due to harvest}
}
\value{
a matrix with 3 rows (non, fine, and coarse woody litter) and 5 columns: a,w,e,n,h corresponding to the amount of carbon in each chemical pool generated from litter each year
}
\description{
Calculate the 'b' Yasso parameter (yearly litter input by chemical pool), given a set of trees with dbh(d), heigh(h) and species(sp)
}
\examples{
trees = data.table(d=c(10,15,20),h=c(10,15,17),sp=c(1,2,2))
get.Yasso.b(trees,turnover=1)

}
\author{
Victor Felix Strîmbu \email{victor.strimbu@nmbu.no}
}
