% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yasso15_wrapper.R
\name{yasso.15}
\alias{yasso.15}
\alias{yasso.15.se}
\title{Yasso15 wrapper}
\usage{
yasso.15(time, climate, init, b, d, steady)

yasso.15.se(time, climate, init, b, d, steady)
}
\arguments{
\item{time}{time interval in years}

\item{climate}{a 3 element vector with climate parameters: mean temperature (°C), accumulated precipitition (mm), and mean difference between min and max monthly temperature(°C)}

\item{init}{5 element vector with initial carbon quantities in each chemical pool}

\item{b}{5 element vector with yearly litter input to each chemical pool}

\item{d}{litter diameter if needed}

\item{steady}{weather to calculate the steady state. If true it ignores "time" parameter and loops until the carbon in chemical pools stabilizes}

\item{time}{time interval in years}

\item{climate}{a 3 element vector with climate parameters: mean temperature (°C), accumulated precipitition (mm), and mean difference between min and max monthly temperature(°C)}

\item{init}{5 element vector with initial carbon quantities in each chemical pool}

\item{b}{5 element vector with yearly litter input to each chemical pool}

\item{d}{litter diameter if needed}

\item{steady}{weather to calculate the steady state. If true it ignores "time" parameter and loops until the carbon in chemical pools stabilizes}
}
\value{
a vector with three elements:

a vector with three elements:
}
\description{
Yasso15 wrapper around the FORTRAN rutine

Calculates the error of soil carbon prediction due to errors in the model parameters
}
\examples{
time = 5
climate = c(10,700,10)
init = c(10,10,10,10,10)
b = c(2,2,2,2,2)
d = 0
steady = F

yasso.15(time,climate,init,b,d,steady)

time = 5
climate = c(10,700,10)
init = c(10,10,10,10,10)
b = c(2,2,2,2,2)
d = 0
steady = F

yasso.15.se(time,climate,init,b,d,steady)

}
\author{
Victor Felix Strîmbu \email{victor.strimbu@nmbu.no}

Victor Felix Strîmbu \email{victor.strimbu@nmbu.no}
}
